forbidden move

+++ db/database/asset.sql.go
@@ -23,7 +23,7 @@ type CreateAssetParams struct {
 	UserID       uuid.UUID
 	Name         string
 	Type         string
-	CurrentValue int32
+	CurrentValue int64
 	PurchasedAt  time.Time
 	CreatedAt    time.Time
 	UpdatedAt    time.Time
