+++ db/schema/20240128151037_create_assets.sql
@@ -5,7 +5,7 @@ CREATE TABLE assets (
     user_id UUID REFERENCES users(id) ON DELETE CASCADE NOT NULL,
     name VARCHAR(255) NOT NULL,
     type VARCHAR(50) NOT NULL,
-    current_value INT NOT NULL,
+    current_value BIGINT NOT NULL,
     purchased_at TIMESTAMP NOT NULL,
     created_at TIMESTAMP NOT NULL,
     updated_at TIMESTAMP NOT NULL
